requirements.txt
ace_tools==0.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
catboost==1.2.5
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cx_Oracle==8.3.0
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.112.2
fastjsonschema==2.20.0
fastremap==1.15.0
filelock==3.15.4
fill_voids==2.0.7
fonttools==4.53.0
fpdf==1.7.2
fqdn==1.5.1
fsspec==2024.6.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
graphviz==0.20.3
h11==0.9.0
h2==3.2.0
hdbscan==0.8.38.post1
hpack==3.0.0
hstspreload==2024.11.1
htmlmin==0.1.12
httpcore==0.9.1
httplib2==0.22.0
humanize==4.10.0
hyperframe==5.2.0
idna==2.10
ImageHash==4.3.1
imageio==2.34.2
imbalanced-learn==0.12.3
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.26.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.4
lightgbm==4.4.0
llvmlite==0.43.0
lungmask @ git+https://github.com/JoHof/lungmask@0653c7019200eb27253c540f731583466fc3c304
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
monai==1.3.2
more-itertools==10.4.0
mpmath==1.3.0
multimethod==1.11.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nilearn==0.10.4
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
optuna==3.6.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.4.0
platformdirs==4.2.2
plotly==5.22.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydicom==2.4.4
PyDrive2==1.21.1
Pygments==2.18.0
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.3.0
pylibjpeg-openjpeg==2.4.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.2.0
shellingham==1.5.4
SimpleITK==2.3.1
six==1.16.0
sklearn-preprocessing==0.1.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.38.4
statsmodels==0.14.2
sweetviz==2.3.1
sympy==1.13.1
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.24
tinycss2==1.3.0
torch==2.4.0
torchaudio==2.4.0
torchio==0.19.9
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typeguard==4.2.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
umap==0.1.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
visions==0.7.6
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
wordcloud==1.9.3
wrapt==1.16.0
xgboost==2.1.1
ydata-profiling==4.8.3
